% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ckron}
\alias{ckron}
\title{Kronecker product}
\usage{
ckron(b, invert = FALSE)
}
\arguments{
\item{b}{List of matrices}

\item{invert}{Logical, take inverse if TRUE}
}
\value{
Matrix with Kronecker product
}
\description{
Repeated Kronecker products on a list of matrices.
}
\author{
Randall Romero-Aguilar, based on Miranda & Fackler's CompEcon toolbox
}
\references{
Miranda, Fackler 2002 Applied Computational Economics and Finance
}

