% Generated by roxygen2 (4.0.2): do not edit by hand
\name{csize}
\alias{csize}
\title{Size on list of matrices}
\usage{
csize(b)
}
\arguments{
\item{b}{List of matrices}
}
\value{
List with fields:
   \itemize{
     \item \code{dimension} scalar, number of matrices
     \item \code{nrows} array, number  of rows per dimension
     \item \code{ncols} array, number  of columns per dimension
   }
}
\description{
Returns dimension information for list of arrays.
}
\author{
Randall Romero-Aguilar, based on Miranda & Fackler's CompEcon toolbox
}
\references{
Miranda, Fackler 2002 Applied Computational Economics and Finance
}

