% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fischer}
\alias{fischer}
\title{Computes Fischer's function and (optionally) its Jacobian}
\usage{
fischer(u, v, du, dv, plus = TRUE)
}
\arguments{
\item{u}{n-array or vector, first function evaluated at x}

\item{v}{n-array or vector, second function evaluated at x}

\item{du}{n-array or vector, derivative of first argument evaluated at x (optional)}

\item{dv}{n-array or vector, derivative of second argument evaluated at x (optional)}

\item{plus}{logical, use 'plus' version if TRUE (default), 'minus' if FALSE (optional)}
}
\value{
If inputs \code{du} and \code{dv} are provided, it returns a list with fields
 \itemize{
     \item \code{f} n.1 matrix, Fischer function  evaluated at x
     \item \code{J} n.n matrix, Jacobian evaluated at x
     }
 Otherwise, it returns only the matrix \code{f}.
}
\description{
Computes Fischer's function and (optionally) its Jacobian
}
\details{
Fischer's function \eqn{\phi:R\times R \rightarrow R}  is specified by
  \deqn{\phi^{\pm}(u,v) = u + v \pm \sqrt{u^2 + v^2}, u,v\in R}
  In this implementation, \code{u,v} are assumed to be function of \code{x}.
}
\examples{
u   <-  c(1,2,3,4)
v   <-  c(1,2,2,1)
du  <-  c(0,1,1,0)
dv  <-  c(1,1,1,1)
fp  <- fischer(u,v,du,dv)
fm  <- fischer(u,v,plus=FALSE)
}
\author{
Randall Romero-Aguilar
}
\references{
Miranda, Fackler 2002 Applied Computational Economics and Finance, section 3.8
}
\seealso{
Other nonlinear equations: \code{\link{minmax}};
  \code{\link{ssmooth}}
}

