% Generated by roxygen2 (4.0.2): do not edit by hand
\name{primes}
\alias{primes}
\title{List of prime numbers}
\usage{
primes(n)
}
\arguments{
\item{n}{integer scalar}
}
\value{
Array with all prime numbers up to \code{n} (inclusive)
}
\description{
List of prime numbers
}
\examples{
primes(100)
}
\author{
Randall Romero-Aguilar
}

