% Generated by roxygen2 (4.0.2): do not edit by hand
\name{qnwlege}
\alias{qnwlege}
\title{Gauss-Legendre quadrature nodes and weights}
\usage{
qnwlege(n, a = matrix(0, 1, length(n)), b = matrix(1, 1, length(n)))
}
\arguments{
\item{n}{1.d number of nodes per dimension}

\item{a}{1.d left endpoints}

\item{b}{1.d right endpoints}
}
\value{
List with fields
 \itemize{
     \item \code{xpoints} prod(n).d quadrature nodes
     \item \code{weights} prod(n).1 quadrature weights
     }
}
\description{
Generates Guass-Legendre quadrature nodes and weights for computing the
definite integral of a real-valued function defined on a hypercube [a,b] in R^d.
}
\examples{
#' # To compute definte integral of a real-valued function f defined on a hypercube
# [a,b] in R^d, write a function f that returns an m.1 vector when passed an
# m.d matrix, and write
q <- qnwlege(n,a,b);
Intf <- crossprod(q$w, f(q$x))

# Alternatively, use the quadrature function
Intf <- quadrature(f,qnwlege,n,a,b)
}
\author{
Randall Romero-Aguilar, based on Miranda & Fackler's CompEcon toolbox
}
\references{
Miranda, Fackler 2002 Applied Computational Economics and Finance
}
\seealso{
Other quadrature functions: \code{\link{qnwequi}};
  \code{\link{qnwlogn}}; \code{\link{qnwnorm}};
  \code{\link{qnwsimp}}; \code{\link{qnwtrap}};
  \code{\link{qnwunif}}; \code{\link{quadrature}}
}
\keyword{quadrature}

