% Generated by roxygen2 (4.0.2): do not edit by hand
\name{qnwlogn}
\alias{qnwlogn}
\title{Gaussian quadrature nodes and weights for lognormal distribution}
\usage{
qnwlogn(n, mu = matrix(0, 1, length(n)), var = diag(length(n)))
}
\arguments{
\item{n}{1.d number of nodes per dimension}

\item{mu}{1.d mean vector (default: zeros)}

\item{var}{d.d positive definite variance matrix (default: identity matrix)}
}
\value{
List with fields
 \itemize{
     \item \code{xpoints} prod(n).d quadrature nodes
     \item \code{weights} prod(n).1 quadrature weights
     }
}
\description{
Generates Gaussian quadrature nodes and probability weights for multivariate
Lognormal distribution with parameters \code{mu} and \code{var}, or,
equivalently, for a random vector whose natural logarithm is normally distributed
 with mean vector \code{mu} and variance matrix \code{var}
}
\examples{
# To compute Ef(X) when f is real-valued and X is lognormal(mu,var) on R^d,
# write a function f that returns m.1 vector when passed an m.d matrix, and write
q  <- qnwlogn(n,mu,var)
Ef  <- crossprod(q$w, f(q$x))

# Alternatively, use the quadrature function
Ef <- quadrature(f,qnwlogn,n,mu,var)
}
\author{
Randall Romero-Aguilar, based on Miranda & Fackler's CompEcon toolbox
}
\references{
Miranda, Fackler 2002 Applied Computational Economics and Finance
}
\seealso{
Other quadrature functions: \code{\link{qnwequi}};
  \code{\link{qnwlege}}; \code{\link{qnwnorm}};
  \code{\link{qnwsimp}}; \code{\link{qnwtrap}};
  \code{\link{qnwunif}}; \code{\link{quadrature}}
}
\keyword{quadrature}

