% Generated by roxygen2 (4.0.2): do not edit by hand
\name{qnwunif}
\alias{qnwunif}
\title{Gaussian quadrature nodes and weights for uniform distribution}
\usage{
qnwunif(n, a = matrix(0, 1, length(n)), b = matrix(1, 1, length(n)))
}
\arguments{
\item{n}{1.d number of nodes per dimension}

\item{a}{1.d left endpoints}

\item{b}{1.d right endpoints}
}
\value{
List with fields
 \itemize{
     \item \code{xpoints} prod(n).d quadrature nodes
     \item \code{weights} prod(n).1 quadrature weights
     }
}
\description{
Generates Gaussian quadrature nodes and probability weights for multivariate
Uniform distribution on hypercube [a,b] in R^d.
}
\examples{
# To compute Ef(X) when f is real-valued and X is Uniform(a,b) on R^d, write a
# function f that returns m.1 vector when passed an m.d matrix, and write
q <- qnwunif(n,a,b)
Ef <- crossprod(q$w,f(q$x))

# Alternatively, use the quadrature function
Ef <- quadrature(f,qnwunif,n,mu,var)
}
\author{
Randall Romero-Aguilar, based on Miranda & Fackler's CompEcon toolbox
}
\references{
Miranda, Fackler 2002 Applied Computational Economics and Finance
}
\seealso{
Other quadrature functions: \code{\link{qnwequi}};
  \code{\link{qnwlege}}; \code{\link{qnwlogn}};
  \code{\link{qnwnorm}}; \code{\link{qnwsimp}};
  \code{\link{qnwtrap}}; \code{\link{quadrature}}
}
\keyword{quadrature}

