% Generated by roxygen2 (4.0.2): do not edit by hand
\name{quadrature}
\alias{quadrature}
\title{Compute numerical quadrature}
\usage{
quadrature(f, method = qnwtrap, n, ...)
}
\arguments{
\item{f}{A function, integrand.}

\item{method}{A function, specifying the quadrature method.}

\item{n}{A scalar, number of nodes.}

\item{...}{Other parameters passed to the selected \code{method}.

  Available \code{method} options are: \code{qnwequi}, \code{qnwlege},
  \code{qnwlogn}, \code{qnwnorm}, \code{qnwsimp}, \code{qnwtrap},
  \code{qnwunif}}
}
\value{
The integral of function \code{f}.
}
\description{
Compute numerical quadrature
}
\examples{
quadrature(function(x) sin(x)^2,qnwtrap,50,0,pi)
}
\author{
Randall Romero-Aguilar
}
\references{
Miranda, Fackler 2002 Applied Computational Economics and Finance, ch.5
}
\seealso{
Other quadrature functions: \code{\link{qnwequi}};
  \code{\link{qnwlege}}; \code{\link{qnwlogn}};
  \code{\link{qnwnorm}}; \code{\link{qnwsimp}};
  \code{\link{qnwtrap}}; \code{\link{qnwunif}}
}
\keyword{quadrature}

