% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ssmooth}
\alias{ssmooth}
\title{Semi-smooth approximation}
\usage{
ssmooth(x, a, b, fx, J)
}
\arguments{
\item{x}{n-array of evaluation points}

\item{a}{n-array of lower bounds}

\item{b}{n-array of upper bounds}

\item{fx}{n-array of function \code{f} evaluated at f}

\item{J}{n.n matrix of Jacobian of \code{f} (optional)}
}
\value{
If input \code{J} is provided, it returns a list with fields
 \itemize{
     \item \code{f} n.1 matrix, function  evaluated at x
     \item \code{J} n.n matrix, Jacobian  evaluated at x
     }
 Otherwise, it returns only the matrix \code{f}.
}
\description{
Computes \eqn{\tilde f(x) = \phi^-(\phi^+(f(x),a-x),b-x)} and its derivative
}
\author{
Randall Romero-Aguilar
}
\references{
Miranda, Fackler 2002 Applied Computational Economics and Finance, section 3.8
}
\seealso{
Other nonlinear equations: \code{\link{fischer}};
  \code{\link{minmax}}
}

