% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vec}
\alias{vec}
\title{VEC operator: stacks matrix columns into a column vector}
\usage{
vec(x)
}
\arguments{
\item{x}{A m.n matrix}
}
\value{
A mn.1 matrix (column vector)
}
\description{
VEC operator: stacks matrix columns into a column vector
}
\examples{
y <- matrix(c(1:6),2,3)
vec(y)
}
\author{
Randall Romero-Aguilar
}

